feat(js-client): extract transport-agnostic core; Node entry rides on it
Extracts RedDB handle, serialization, auth-merge, login, insert-id
normalization, embedded-URI rejection, and NDJSON frame helpers into
src/core/ (no node: imports). Node entry consumes the core and injects
the existing node:stream streaming impl; public API unchanged.
129 js-client tests pass (127 ok, 2 skip).

Closes #875
